lesson_5/blog_project/requirements.txt
alembic==1.7.7
asn1crypto==1.5.1
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.2
blinker==1.4
cffi==1.15.0
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.4.6
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
pg8000==1.26.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pyparsing==3.0.8
pytest==7.1.1
python-dateutil==2.8.2
python-decouple==3.6
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
